requirements.txt
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10
gunicorn==19.4.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-decouple==3.1
scrapy-djangoitem==1.1.1
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
traitlets==4.3.2
wcwidth==0.1.7
